pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "polybot"
version = "0.1.0"
description = "Polymarket prediction market arbitrage trading bot"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "py-clob-client>=0.34.0",
    "aiohttp>=3.9.0",
    "websockets>=12.0",
    "web3>=6.15.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "orjson>=3.9.0",
    "uvloop>=0.19.0",
    "streamlit>=1.29.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "eth-account>=0.10.0",
    "loguru>=0.7.0",
    "aiosqlite>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.9",
]

[project.scripts]
polybot = "polybot.main:main"
polybot-dashboard = "polybot.dashboard.app:run"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
